requirements.txt
python-telegram-bot==20.7
yt-dlp==2024.1.4
requests==2.31.0
python-dotenv==1.0.0
